+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+# next
+
+* added observeMutations() method to behaviors (#22)
+* conditions (responsive af)!
+
 # 0.0.1
 
 Nothing, duh.